Add test check find by serial number in configuration service
and minor correction in VO migrate to Long

+++ registry/src/main/java/com/gl/registry/vo/DeviceResponseVO.java
@@ -18,7 +18,7 @@ import java.net.InetAddress;
 public class DeviceResponseVO extends DeviceBaseVO {
 
     @NotNull
-    private Integer deviceId;
+    private Long deviceId;
 
     @NotBlank
     @Size(min = 7, max = 15)
